deploy script

+++ target_chains/ethereum/examples/coin_flip/contract/src/CoinFlip.sol
@@ -54,7 +54,7 @@ contract CoinFlip {
             userRandom,
             providerRandom
         );
-        emit FlipResult(randomNumber % 2 == 0);
+        emit FlipResult(uint256(randomNumber) % 2 == 0);
     }
 
     // Reinitialize the parameters of this contract.
